+++ docs/RELEASE-NOTES.md
@@ -1,13 +1,19 @@
 # Release Notes
 
-## 1.1.0 - ?
+## 1.1.0 - 08 January 2021
 
 Breaking changes:
 - (None)
 
 New features:
+- Added `EnvironmentVariableProvider.GetByte`
+- Added `EnvironmentVariableProvider.GetByteOrDefault`
+- Added `EnvironmentVariableProvider.GetShort`
+- Added `EnvironmentVariableProvider.GetShortOrDefault`
 - Added `EnvironmentVariableProvider.GetLong`
 - Added `EnvironmentVariableProvider.GetLongOrDefault`
+- Added `EnvironmentVariableProvider.GetFloat`
+- Added `EnvironmentVariableProvider.GetFloatOrDefault`
 - Added `EnvironmentVariableProvider.GetDouble`
 - Added `EnvironmentVariableProvider.GetDoubleOrDefault`
 - Added `EnvironmentVariableProvider.GetDecimal`